firmware-imx: use undepended copies for FW-packages
In order to have high compatibility with newer meta-freescale layers,
but still maintain deterministic NBOOT builds, the firmware packages are
frozen in meta-fus-nboot.

When a new version of the firmware is released, it must be frozen for
future NBOOT builds.

The selection of firmware versions, including the NBOOT version to
build, must be defined in the machine configuration. This is done using
PREFERRED_VERSION_${BPN} = "${PV}".

Examples can be found in fsimx93.conf or fsimx8ulp.conf.